refactor(scene): walking_position re-homes to motion — sim->render DAG is clean
The pure Point lerp lived in pixel_painter/anchors.rs while
pose/mod.rs's prod pose-history recording imported it — the ONE
back-edge keeping sim->render from being a clean DAG inside the scene
crate (and the blocker for the gamma-wave sim/paint split). Verbatim
move to motion (beside octile_path_len); anchors re-imports; the stale
"renderer.rs applies vertical breathing" comment replaced with an
accurate doc. gen-media --check: zero pixel drift.

Edge inventory for the split (recorded): pose/motion/pathfind now have
ZERO prod pixel_painter imports; only 7 test-only character_anchor
uses remain in pose/tests.rs.
