+++ crates/pixtuoid-scene/src/motion/mod.rs
@@ -604,5 +604,23 @@ pub(crate) fn settle_len(approach: Point, seat: Option<Point>) -> u32 {
     seat.map_or(0, |s| octile_distance(approach, s))
 }
 
+/// Pure linear interpolation along the walk segment `from → to` at
+/// `t_x1000` (0..=1000). Deterministic walk-leg geometry: the pose history
+/// records with it (snap-back lookups need the breath-free position) and
+/// `pixel_painter` re-imports it to place the walking sprite/label anchors.
+pub(crate) fn walking_position(from: Point, to: Point, t_x1000: u16) -> Point {
+    let t = t_x1000 as i32;
+    let dx = to.x as i32 - from.x as i32;
+    let dy = to.y as i32 - from.y as i32;
+    // Clamp at zero before casting to u16 — left-walking agents (to.x <
+    // from.x) cross through negative x partway through their walk if the
+    // animation interpolation overshoots, and a bare `as u16` cast wraps
+    // silently to ~65k, blitting the sprite off-screen invisibly.
+    Point {
+        x: (from.x as i32 + dx * t / 1000).max(0).min(u16::MAX as i32) as u16,
+        y: (from.y as i32 + dy * t / 1000).max(0).min(u16::MAX as i32) as u16,
+    }
+}
+
 #[cfg(test)]
 mod tests;
